Fixed response code for list results with more results and blocking loading of routes when the user should not be able to access the route

+++ rest/dto/src/main/scala/com/jshnd/assassin/dto/ListResult.scala
@@ -23,7 +23,7 @@ case class ListResult[T](@xmlElementWrapper(name = "items")
 
                          @xmlElement
                          @BeanProperty
-                         var totalRecords: Int) {
+                         var totalRecords: Long) {
 
   def this() = this(null, 0, 0, 0)
